__io_uring_get_cqe: silence signed vs unsigned comparison warning
Some compilers throw:

file: "/libs/liburing/src/queue.c"
line: 61 warning: comparison of integer expressions of different
signedness: ‘int’ and ‘unsigned int’
`} else if (ret == submit) {`

Shouldn't really matter, we already checked if ret < 0 before doing
this comparison. And since 'submit' or 'ret' can't be larger than the
max ring size, there's on risk of overflow either.

But just cast the 'submit' variable to a signed int, that should get
rid of it.

+++ src/queue.c
@@ -58,7 +58,7 @@ int __io_uring_get_cqe(struct io_uring *ring, struct io_uring_cqe **cqe_ptr,
 						   wait_nr, flags, sigmask);
 		if (ret < 0) {
 			err = -errno;
-		} else if (ret == submit) {
+		} else if (ret == (int)submit) {
 			submit = 0;
 			wait_nr = 0;
 		} else {
